伤心秦汉经行处，宫阙万间都做了土
ai评分 增加一个随机量，内景可触发事件，计算鼠标对应坐标略修改

+++ src/BattleMenu.h
@@ -32,7 +32,7 @@ public:
     struct AIAction
     {
         int Action;
-        int point = 0;
+        double point = 0;
         int MoveX, MoveY;
         int ActionX, ActionY;
         Magic* magic = nullptr;
